+++ objectified-ui/public/WHATS_NEW.md
@@ -104,6 +104,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Adds dry run support to test imports before importing data
   - Adds incremental mode that will import all available, but skip failures
   - Failed import items are listed in red
+  - Skipped import items are listed in gray
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions
   - Improved import support for paths and securitySchemes